pyproject.toml
[project]
name = "terminal-audio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keyboard>=0.13.5",
    "python-vlc>=3.0.21203",
    "rich>=14.2.0",
    "yt-dlp>=2025.12.8",
    "ytmusicapi>=1.11.4",
]
